refactor(llamabot)🔧: Remove sqlite-vec dependency and related code
- Removed sqlite-vec from dependencies in pyproject.toml and pixi.lock.
- Deleted SQLiteVecDocStore class from docstore.py and archived its implementation.
- Removed sqlite-vec package entries from pixi.lock.
