+++ pyproject.toml
@@ -102,7 +102,7 @@ dependencies = [
     "fastapi",
     "uvicorn",
     "sentence-transformers",
-    "tenacity", "sqlite-vec>=0.1.3,<0.2",
+    "tenacity",
 ]
 requires-python = "<3.13"
 description = "A Pythonic interface to LLMs."
